removed sync_record
