+++ osvauld-db/src/models.rs
@@ -10,11 +10,6 @@ use osvauld_core::models::{
     resource::Resource as DomainResource,
     resource_key::ResourceKey as DomainResourceKey,
     share_record::{PermissionLevel, ShareOperation, ShareRecord as DomainShareRecord},
-    sync_record::{
-        DeviceRecord as DomainDeviceRecord, DeviceRecordStatus as DomainDeviceRecordStatus,
-        SyncRecord as DomainSyncRecord,
-    },
-    sync_types::{OperationType, ResourceType, SyncStatus},
     user::User as DomainUser,
     vector_clock::ResourceVectorClock as DomainResourceVectorClock,
 };
@@ -72,111 +67,6 @@ pub struct SyncRecordModel {
     pub source_device_id: String,
     pub created_at: i64,
     pub updated_at: i64,
-}
-
-impl SyncRecordModel {
-    pub fn to_domain(&self) -> DomainSyncRecord {
-        DomainSyncRecord {
-            id: self.id.clone(),
-            resource_id: self.resource_id.clone(),
-            resource_type: ResourceType::from(self.resource_type.clone()),
-            operation_type: OperationType::from(self.operation_type.clone()),
-            source_device_id: self.source_device_id.clone(),
-            created_at: self.created_at,
-            updated_at: self.updated_at,
-        }
-    }
-}
-
-impl From<&DomainSyncRecord> for SyncRecordModel {
-    fn from(record: &DomainSyncRecord) -> Self {
-        Self {
-            id: record.id.clone(),
-            resource_id: record.resource_id.clone(),
-            resource_type: record.resource_type.to_string(),
-            operation_type: record.operation_type.to_string(),
-            source_device_id: record.source_device_id.clone(),
-            created_at: record.created_at,
-            updated_at: record.updated_at,
-        }
-    }
-}
-
-#[derive(Queryable, Insertable, Selectable, Debug)]
-#[diesel(table_name = device_records)]
-pub struct DeviceRecordModel {
-    pub id: String,
-    pub sync_record_id: String,
-    pub device_id: String,
-    pub status: String,
-    pub synced: bool,
-    pub created_at: i64,
-    pub updated_at: i64,
-}
-
-impl DeviceRecordModel {
-    pub fn to_domain(&self) -> DomainDeviceRecord {
-        DomainDeviceRecord {
-            id: self.id.clone(),
-            sync_record_id: self.sync_record_id.clone(),
-            device_id: self.device_id.clone(),
-            status: SyncStatus::from(self.status.clone()),
-            synced: self.synced,
-            created_at: self.created_at,
-            updated_at: self.updated_at,
-        }
-    }
-}
-
-impl From<&DomainDeviceRecord> for DeviceRecordModel {
-    fn from(record: &DomainDeviceRecord) -> Self {
-        Self {
-            id: record.id.clone(),
-            sync_record_id: record.sync_record_id.clone(),
-            device_id: record.device_id.clone(),
-            status: record.status.to_string(),
-            synced: record.synced,
-            created_at: record.created_at,
-            updated_at: record.updated_at,
-        }
-    }
-}
-
-#[derive(Queryable, Insertable, Selectable, Debug)]
-#[diesel(table_name = device_record_status)]
-pub struct DeviceRecordStatusModel {
-    pub id: String,
-    pub device_record_id: String,
-    pub aware_device_id: String,
-    pub synced: bool,
-    pub created_at: i64,
-    pub updated_at: i64,
-}
-
-impl DeviceRecordStatusModel {
-    pub fn to_domain(&self) -> DomainDeviceRecordStatus {
-        DomainDeviceRecordStatus {
-            id: self.id.clone(),
-            device_record_id: self.device_record_id.clone(),
-            aware_device_id: self.aware_device_id.clone(),
-            synced: self.synced,
-            created_at: self.created_at,
-            updated_at: self.updated_at,
-        }
-    }
-}
-
-impl From<&DomainDeviceRecordStatus> for DeviceRecordStatusModel {
-    fn from(status: &DomainDeviceRecordStatus) -> Self {
-        Self {
-            id: status.id.clone(),
-            device_record_id: status.device_record_id.clone(),
-            aware_device_id: status.aware_device_id.clone(),
-            synced: status.synced,
-            created_at: status.created_at,
-            updated_at: status.updated_at,
-        }
-    }
 }
 
 #[derive(Queryable, Insertable, Identifiable, Selectable)]
