+++ osvauld-services/src/lib.rs
@@ -1,13 +1,11 @@
 // lib.rs
 mod auth_service2;
-mod folder_service;
 mod folder_service2;
 mod resource_service2;
 mod sync_service2;
 mod user_service2;
 
 pub use auth_service2::*;
-pub use folder_service::FolderService;
 pub use folder_service2::*;
 pub use resource_service2::*;
 pub use sync_service2::*;
